backend/requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
groq
python-dotenv==1.0.1
numpy
onnxruntime
pooch
tqdm
rembg